improve add expense and create helper
