+++ api/controllers/wallet.controller.js
@@ -10,41 +10,11 @@ const {
 	BAD_REQUEST_STATUS,
 } = require('../constants/httpStatus.constant');
 
-const getTotalVirtualWallet = (wallets) => {
-	const total = wallets.reduce(
-		(currentTotal, wallet) => currentTotal + wallet.accountBalance,
-		0
-	);
-
-	return total;
-};
-
-const updateVirtualTransactions = (transactions, transaction) => {
-	const now = moment(transaction.date).format('MMMM Do YYYY');
-
-	const transactionIndex = transactions.findIndex(
-		(obj) => moment(obj.date).format('MMMM Do YYYY') === now
-	);
-
-	if (transactionIndex !== -1) {
-		transactions[transactionIndex].expenses = transactions[
-			transactionIndex
-		].expenses.concat(transaction.expenses);
-	} else {
-		transactions.push(transaction);
-	}
-};
-
-const getTransactionsVirtualWallet = (wallets) => {
-	const transactions = wallets.reduce((currentTransactions, wallet) => {
-		wallet.transactions.forEach((transaction) => {
-			updateVirtualTransactions(currentTransactions, transaction);
-		});
-		return currentTransactions;
-	}, []);
-
-	return transactions;
-};
+const {
+	getTotalVirtualWallet,
+	updateVirtualTransactions,
+	getTransactionsVirtualWallet,
+} = require('../helper/helper');
 
 module.exports.index = async (req, res) => {
 	const { _id } = req.user;
